requirements.txt
python-dotenv
selenium
webdriver_manager
oauth2client
gspread